DuckDBRowReader: Add class to stream DuckDB results efficiently
There currently exists the `DuckDBResultReader`, but the issue that
we're running into is that this processes chunks into an internal buffer
that's never cleared. For large result sets, this buffer builds and
builds until we run out of memory.

The case that we have is that we just need the result set row by row
without needing to compute any of the column data. This new class does
less than `DuckDBResultReader` by not storing the results of the query
in the underlying buffer. Instead, the results of the chunk are passed
back to the caller where the caller can handle the chunk as they wish.

The interface for processing rows is similar to `DuckDBRowReader` where
the caller can still use converters as they wish to process their data.